enhancement(aws_sqs source): Add additional config options
Mirroring the options on the `aws_s3` source:

* `delete_message` to control whether messages are deleted after
  processing. This is useful for testing out the source.
* `visibility_timeout_secs` to control how long messages are locked for
  before being rereleased to be processed again. Tuning this is useful
  for controlling how long a message will be "sent" if a Vector instance
  crashes before deleting the message.

Closes: #11927

+++ src/sources/aws_sqs/source.rs
@@ -32,6 +32,8 @@ pub struct SqsSource {
     pub queue_url: String,
     pub decoder: Decoder,
     pub poll_secs: u32,
+    pub visibility_timeout_secs: u32,
+    pub delete_message: bool,
     pub concurrency: u32,
     pub(super) acknowledgements: bool,
 }
@@ -74,6 +76,7 @@ impl SqsSource {
             .queue_url(&self.queue_url)
             .max_number_of_messages(MAX_BATCH_SIZE)
             .wait_time_seconds(self.poll_secs as i32)
+            .visibility_timeout(self.visibility_timeout_secs as i32)
             // I think this should be a known attribute
             // https://github.com/awslabs/aws-sdk-rust/issues/411
             .attribute_names(QueueAttributeName::Unknown(String::from("SentTimestamp")))
@@ -120,17 +123,19 @@ impl SqsSource {
                 emit!(StreamClosedError { error, count });
             }
 
-            if let Some(receiver) = batch_receiver {
-                let client = self.client.clone();
-                let queue_url = self.queue_url.clone();
-                tokio::spawn(async move {
-                    let batch_status = receiver.await;
-                    if batch_status == BatchStatus::Delivered {
-                        delete_messages(&client, &receipts_to_ack, &queue_url).await;
-                    }
-                });
-            } else {
-                delete_messages(&self.client, &receipts_to_ack, &self.queue_url).await;
+            if self.delete_message {
+                if let Some(receiver) = batch_receiver {
+                    let client = self.client.clone();
+                    let queue_url = self.queue_url.clone();
+                    tokio::spawn(async move {
+                        let batch_status = receiver.await;
+                        if batch_status == BatchStatus::Delivered {
+                            delete_messages(&client, &receipts_to_ack, &queue_url).await;
+                        }
+                    });
+                } else {
+                    delete_messages(&self.client, &receipts_to_ack, &self.queue_url).await;
+                }
             }
         }
     }
